requirements.txt
anyio==3.6.2
azure-functions==1.12.0
click==8.1.3
fastapi==0.89.1
h11==0.14.0
idna==3.4
joblib==1.2.0
numpy==1.24.1
pydantic==1.10.4
scikit-learn==1.2.0
scipy==1.10.0
sniffio==1.3.0
starlette==0.22.0
threadpoolctl==3.1.0
typing-extensions==4.4.0
uvicorn==0.20.0
